Compile time error when a uint8 is used as a FBlackboard::Fkey Fix BlackboardDecoratorDetails and BlackboardSelectorDetails #tests Test FBlackboardDecoratorDetails in editor, rebuild projects #rb mieszko.zielinski, maxime.mercier
[CL 23020972 by charles lefebvre in ue5-main branch]

+++ Engine/Source/Editor/BehaviorTreeEditor/Private/DetailCustomizations/BlackboardDecoratorDetails.cpp
@@ -138,7 +138,7 @@ void FBlackboardDecoratorDetails::OnKeyIDChanged()
 		return;
 	}
 
-	uint8 KeyID;
+	int32 KeyID;
 	FPropertyAccess::Result Result = KeyIDProperty->GetValue(KeyID);
 	if (Result == FPropertyAccess::Success)
 	{
